backend/requirements.txt
Flask>=3.0
Flask-CORS>=4.0
Flask-JWT-Extended>=4.6
Flask-SQLAlchemy>=3.1
gunicorn>=22.0
Werkzeug>=3.0
uvicorn
